Re-work the determining of what we were called as
Previously we were checking argv[0][0] to see if it was a '/' or a '.'.
If so, we used rindex() to strip off all the path, otherwise we just
used argv[0].

Just checking argv[0][0] for a '/' isn't technically enough as we could
get called like; sbin/recmail. So, do away with this check altogether.
Instead, use strstr() to see if there is a '/' in the path, if so use
use strrchr()[0] to strip of the path otherwise just use argv[0].

[0] - rindex() is marked as LEGACY in POSIX.1-2001 and POSIX.1-2008
      removes it.

Signed-off-by: Andrew Clayton <andrew@pccl.info>